Little refactoring written before code
